fixing pandas_storage.py

+++ urls_crowler/dto/result_message_dto.py
@@ -4,4 +4,4 @@ from pydantic import BaseModel
 class ResultMessageDTO(BaseModel):
     all_links_count: int
     time_spent: str
-    av_speed: float
+    av_speed: int
